refactor: move drawing menu to separate component
